+++ .github/workflows/deploy.yml
@@ -51,8 +51,8 @@ jobs:
         uses: jacopocarlini/azure-pipelines@v1.3
         with:
           azure-devops-project-url: https://dev.azure.com/pagopaspa/pagoPA-projects
-          azure-pipeline-name: 'projectName.deploy' # TODO: set the name
-          azure-devops-token: ${{ secrets.AZURE_DEVOPS_TOKEN }} # TODO: set this secret in repository
+          azure-pipeline-name: 'pagopa-biz-events.deploy'
+          azure-devops-token: ${{ secrets.AZURE_DEVOPS_TOKEN }}
           azure-template-parameters: '{"ENV": "dev", "SEMVER": "${{env.SEMVER}}", "TEST": "true", "FORCE_REPLACE_DOCKER_IMAGE": "false"}'
           azure-pipeline-variables: '{"system.debug": "true"}'
