+++ CODEOWNERS
@@ -1,4 +1,4 @@
 # see https://help.github.com/en/articles/about-code-owners#example-of-a-codeowners-file
 
 * @pagopa/pagopa-tech
-# TODO: set your codeowners
+
