+++ openapi/generate_openapi.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
-curl http://localhost:8080/v3/api-docs | python3 -m json.tool > ./openapi.json
+curl http://localhost:8687/v3/api-docs | python3 -m json.tool > ./openapi.json
 
 # UI mode http://localhost:8080/swagger-ui/index.html
