+++ src/main/java/it/gov/pagopa/bizeventsservice/Application.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.microservice; // TODO: refactor the package
+package it.gov.pagopa.bizeventsservice;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
