+++ src/main/java/it/gov/pagopa/bizeventsservice/config/MappingsConfiguration.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.microservice.config;
+package it.gov.pagopa.bizeventsservice.config;
 
 
 import org.modelmapper.ModelMapper;
