+++ src/main/java/it/gov/pagopa/bizeventsservice/config/OpenApiConfig.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.microservice.config;
+package it.gov.pagopa.bizeventsservice.config;
 
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
@@ -19,7 +19,7 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
 
-import static it.gov.pagopa.microservice.util.Constants.HEADER_REQUEST_ID;
+import static it.gov.pagopa.bizeventsservice.util.Constants.HEADER_REQUEST_ID;
 
 @Configuration
 public class OpenApiConfig {
