+++ src/main/java/it/gov/pagopa/bizeventsservice/config/WebMvcConfiguration.java
@@ -1,7 +1,7 @@
-package it.gov.pagopa.microservice.config;
+package it.gov.pagopa.bizeventsservice.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import it.gov.pagopa.microservice.model.AppCorsConfiguration;
+import it.gov.pagopa.bizeventsservice.model.AppCorsConfiguration;
 import lombok.SneakyThrows;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
