+++ src/main/java/it/gov/pagopa/bizeventsservice/controller/HomeController.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.microservice.controller;
+package it.gov.pagopa.bizeventsservice.controller;
 
 import io.swagger.v3.oas.annotations.Hidden;
 import org.springframework.beans.factory.annotation.Value;
