+++ src/main/java/it/gov/pagopa/bizeventsservice/exception/AppError.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.microservice.exception;
+package it.gov.pagopa.bizeventsservice.exception;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
