+++ src/main/resources/application-local.properties
@@ -1,6 +1,8 @@
 # Info
 info.properties.environment=local
 
+server.port=8687
+
 # Logging
 logging.level.root=INFO
 logging.level.it.gov.pagopa=DEBUG
